feat(coordinator): integrate coordinator module and enhance employee management
- Added the `coordinator` module as a new dependency in the project, updating both `Cargo.toml` and `Cargo.lock`.
- Modified the `LlmAdapter` trait to accept a `RunRecord` instead of a `String`, improving type safety.
- Introduced new API routes for employee termination, enhancing employee management capabilities.
- Updated the `EmployeeModel` and `EmployeeRecord` structures to include a `last_run_at` field for better tracking of employee activity.
- Enhanced the `EmployeeRepository` with methods to manage employee records and their associated runs.
- Streamlined the project structure by integrating the new `runs` module into the API routes.

+++ crates/persistence/src/models_v2/employees/types.rs
@@ -127,7 +127,7 @@ impl FromStr for EmployeeRole {
   }
 }
 
-#[derive(Clone, Debug, serde::Serialize, serde::Deserialize, SurrealEnumValue)]
+#[derive(Clone, Debug, PartialEq, Eq, serde::Serialize, serde::Deserialize, SurrealEnumValue)]
 #[serde(rename_all = "snake_case")]
 pub enum EmployeeStatus {
   Idle,
@@ -172,10 +172,10 @@ impl Default for EmployeeProviderConfig {
 
 #[derive(Clone, Default, Debug, serde::Serialize, serde::Deserialize, SurrealValue)]
 pub struct EmployeeRuntimeConfig {
-  pub heartbeat_interval_sec: i32,
+  pub heartbeat_interval_sec: i64,
   pub heartbeat_prompt:       String,
   pub wake_on_demand:         bool,
-  pub max_concurrent_runs:    i32,
+  pub max_concurrent_runs:    i64,
 }
 
 #[derive(Clone, Default, Debug, serde::Serialize, serde::Deserialize, SurrealValue)]
